Added an aws.mime file to Yolk.
The aws.mime file is actually just a copy of the Apache mime.types file.
This file contains a boat-load of MIME types. According to the comment
found in aws-mime.ads, AWS will only load these new MIME types if the
file is named aws.mime and it is placed in the same folder as the
executable.

	new file:   exe/aws.mime