Got rid of Is_Empty and added some comments.
I had a bunch of Is_Empty calls in simple_email.adb as a part of the
exception system, but I've decided to get rid of those and instead
raise exceptions in the Send procedure.

I've also started to write comments for the Simple_Email package. This
was long overdue.

	modified:   core/simple_email.adb
	modified:   core/simple_email.ads