Really horrible commit with _lots_ of changes.
This commit is a mess. I've started work on the new DB access system
where the GNATCOLL.SQL connections are generated in a set of new tasks.

The goal is to be able to have several connections to different databases
all working in the same AWS thread.

Currently it's pretty messy and very raw.

	modified:   README
	modified:   core/config_file_parser.adb
	modified:   core/config_file_parser.ads
	modified:   core/configuration.ads
	new file:   core/connect_to_db-postgresql.adb
	new file:   core/connect_to_db-postgresql.ads
	new file:   core/connect_to_db.adb
	new file:   core/connect_to_db.ads
	modified:   core/db_connection.adb
	modified:   core/db_connection.ads
	modified:   core/simple_email.adb
	modified:   core/simple_email.ads
	modified:   exe/aws.ini
	modified:   exe/configuration/GNATCOLL.SQL.Logs.ini
	modified:   my_view/view-index.adb
	modified:   my_view/view.ads
	modified:   yolk.adb
	modified:   yolk.gpr