Adding task-to-task association to the database system.
I want to have a 1 to 1 correlation between an AWS thread and a GNATCOLL.SQL
database connection object. If AWS thread A requests a database connection,
then it should _always_ get the connection started by thread 1. If B requests
one, it should always get the one started by 2, and so on.

To accomplish this I've added a map to the connect_to_db-postgresql generic.

This map is encapsulated in a protected object, where multiple readers are
allowed, but only one writer.

It should only lock due to writers when a new DB_Conn object is created and
set. This happens as many times as there are AWS threads.

There are still some weaknesses in the system: It doesn't handle similar
hashes yet, which obviously is a bad thing.

	modified:   core/connect_to_db-postgresql.adb
	modified:   core/connect_to_db-postgresql.ads
	modified:   core/process_control.adb
	modified:   exe/aws.ini
	modified:   my_view/view-index.adb