Slimmed down the connect_to_db-postgresql package.
Seeing as it is a child of connect_to_db, why not just keep all the core
functionality in the parent package, and keep the generic package as slim
as possible.

Doing this does not appear to hurt performance, but it does make it a lot easier
to add a new DBMS generic to the mix.

	modified:   core/connect_to_db-postgresql.adb
	copied:     core/connect_to_db.ads -> core/connect_to_db-postgresql.ads
	modified:   core/connect_to_db.adb
	copied:     core/connect_to_db-postgresql.ads -> core/connect_to_db.ads