Added a logfile clean up system.
There's a new task in yolk.adb that takes care of cleaning up old log files.

It simply polls the directory where the AWS access and error logs are placed,
and deletes the oldest log files if there are more than X files.

How often the task polls is still undecided, as is how many log files it will
keep intact. I'm thinking something along the lines of once on the hour, and 30
files.

	new file:   core/logfile_cleanup.adb
	new file:   core/logfile_cleanup.ads
	modified:   yolk.adb