Added support for database connections using AWS and/or DB_Conn tasks.
The old Connect_To_DB and Connect_To_DB.PostgreSQL packages only supported
connecting to databases using the Connect_To_DB.DB_Conn tasks. These were
then mapped to the AWS tasks using a hashed map.

This makes sense in scenarios where multiple database connections are
necessary in a single AWS task, but it makes no sense at all if only one
database connection is needed.

So to get rid of the hashed map overhead, I've added a Connection_Mapping_Method
type to Connect_To_DB and a Task_To_DB_Mapping_Method : Connection_Mapping_Method
parameter to Connect_To_DB.PostgreSQL. This parameter specify whether we wish to
connect to the database using the AWS tasks, or if we want to use the DB_Conn
tasks.

It is possible to mix the two, so one (the primary) connection is using AWS
tasks and others connections are using DB_Conn.

Examples provided in the source.

	modified:   core/connect_to_db-postgresql.adb
	modified:   core/connect_to_db-postgresql.ads
	modified:   core/connect_to_db.adb
	modified:   core/connect_to_db.ads
	modified:   my_view/view-index.adb
	modified:   my_view/view.ads