Added the AWS.MIME.Load procedure call.
This is a work-around for loading the aws.mime file in the configuration
directory.

The AWS.MIME package initializes itself with the default configuration
parameters because it is elaborated before we can set our configuration
in the Configuration package. This means that the path to the aws.mime
file is the default path, which is ./

This obviously wont work.

With this call we can "lazily" load a MIME file after the fact.

	modified:   yolk.adb