Got rid of the unknown content handler and registed a default callback instead.
This also prompted moving the Register_Default_Callback call to the
My_Handlers package, instead of letting it sit in Handlers. The reason for
this change is rather simple: The user should not have to mess around in a
core file just to change the default callback to something else than what is
the default Yolk behavior.

	modified:   core/configuration.ads
	modified:   core/handlers.adb
	modified:   exe/configuration/config.ini
	modified:   my_handlers/my_handlers.adb
	modified:   my_handlers/my_handlers.ads