Got rid of the hashed map and replaced it with Ada.Task_Attributes.
Instead of mapping the AWS threads to the DB_Conn tasks using a hashed map
(with hashes based on the string image of the task ID), I've switched to
using Ada.Task_Attributes instead. With this package it is possible to
register attributes to a specific task using a generic.

	modified:   core/connect_to_db-postgresql.adb
	modified:   core/connect_to_db-postgresql.ads
	modified:   core/connect_to_db.adb
	modified:   core/connect_to_db.ads