Moved all the my_* files to the new demo/ directory.
We really shouldn't have any my_* files tracked in their actual directories, as
we'd like to avoid overwriting them on an update of Yolk.

The added "hassle" of this is that a new clone of Yolk now requires some manual
copying of files before it will compile. This is a cost worth paying.