Changes to Yolk.Static_Content. Renamed functions and added check for age.
Instead of the old Binary_File and Text_File functions, we now have
Non_Compressable and Compressable.

Also it's now possible the define a max age on the compressed .gz files.
If the modification time of a .gz file is older than the set max age, then
the file is deleted and created a new.

These changes also meant adding an extra configuration parameter, and
doing some function name adjustments to Yolk.Handlers.