Rework. Atom_Entry_Source is now a limited private type.
If you need to add the atom:source element to an entry, you instantiate a source
object and add to an entry, just like you add entries to the feed.

Created Null_Atom_Entry and Null_Atom_Entry_Source constant in the spec, so
things are a bit cleaner in the body.